editoast: move sql request from views::infra::edition to modelsv2::infra
